Move `nuxt-ts` logic to `@nuxt/typescript`
