+++ distributions/nuxt-ts/package.js
@@ -1,7 +1,5 @@
-import { readJSON, writeJSON } from 'fs-extra'
-
 export default {
-  build: true,
+  build: false,
   hooks: {
     async 'build:done'(pkg) {
       const mono = pkg.load('../..')
@@ -20,17 +18,6 @@ export default {
       ])
 
       await pkg.writePackage()
-
-      if (pkg.options.suffix && pkg.options.linkedDependencies) {
-        const tsconfig = await readJSON(pkg.resolvePath('tsconfig.json'))
-
-        tsconfig.compilerOptions.types = tsconfig.compilerOptions.types.map((type) => {
-          const suffix = pkg.options.linkedDependencies.includes(type) ? pkg.options.suffix : ''
-          return type + suffix
-        })
-
-        await writeJSON(pkg.resolvePath('tsconfig.json'), tsconfig, { spaces: 2 })
-      }
     }
   }
 }
