+++ packages/typescript/src/ts-node.js
@@ -1,0 +1,6 @@
+import { register } from 'ts-node'
+
+export function registerTsNode() {
+  // https://github.com/TypeStrong/ts-node
+  register()
+}
